Agent007_v2/requirements.txt
kivymd==1.1.1
kivy==2.3.0
google-generativeai
python-dotenv
